update all frontend

+++ backend_ai_service/app/api/jobs.py
@@ -1,11 +1,14 @@
+
 import json
 import logging
 import uuid
+from decimal import Decimal
 from typing import Optional, List
 from datetime import datetime, timedelta, timezone
 
 from fastapi import APIRouter, HTTPException, BackgroundTasks
 from pydantic import BaseModel
+
 from app.core.supabase import supabase
 from app.core.gemini import gemini
 from app.prompts.job_matching import JOB_MATCHING_PROMPT
@@ -22,8 +25,8 @@ class JobCreateRequest(BaseModel):
     description: str
     requirements: Optional[str] = None
     location: Optional[str] = None
-    salary_from: Optional[int] = 0
-    salary_to: Optional[int] = 0
+    salary_from: Optional[Decimal] = 0.0
+    salary_to: Optional[Decimal] = 0.0
     salary_unit: Optional[str] = 'month'
     currency: Optional[str] = 'VND'
     expired_at: Optional[str] = None
